requirements.txt
anyio==3.6.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
pydantic==1.9.1
pymongo==4.2.0
PyMySQL==1.0.2
python-decouple==3.6
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.20.4
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
